requirements.txt
pytz==2022.1
requests==2.28.1
python-dotenv~=0.20.0
ipython
pandas
numpy
tqdm
tinkoff-investments
tinkoff

# Django
django
django-timezone-field 
django-cors-headers
django-debug-toolbar
whitenoise

# Django 3.0 async requirements
gunicorn
uvicorn
httptools

# Databases
psycopg2-binary
dj-database-url

# Distributed tasks
celery
redis
django-celery-beat

# Telegram
python-telegram-bot

# monitoring
# sentry-sdk

# Django 3.0 async requirements
uvloop

asgiref~=3.5.2
python-dateutil~=2.8.2
toolz